chore: Migrate to npm scripts from grunt
Note that not all functionality has been ported to npm scripts. Therefore
the following grunt tasks are missing:
- running tests on browserstack
- removing logging
- producing a bundle in lib

`npm start`:
This uses budo to start a dev server that:
- opens browser automatically
- live reloads on js/html changes
- transforms based on .babelrc file

Note that this does not output anything to dist/, but simply serves
the browser/babelified output from requests to dist/finput.js.

chore: Update readme to deal with removal of grunt in favour of npm