chore: Add lib build (distinct from dist build)
dist:
For use in html script tags. Including this script makes a global finput
object available. This is a bundle that includes all of finput and its
dependencies.

lib:
For use in npm package imports/requires. This is **not** a bundle and
includes all of finput's transpiled source files so that they work in
a node environment. The idea is that anyone installing and requiring
finput will then use a bundler on their entry file which will bundle
finput in with the rest of their dependencies.