chore: Remove engines from package.json
Those versions of node/npm were required for development only. `npm` can
be told to check the engines when installing and bail if they are
incorrect. This is bad for anyone installing our package as they may
not be using the listed version. Instead we rely on CI choosing correct
node/npm from .nvmrc and remove from package.json.